added functionality for the view author button

+++ components/BookCard.js
@@ -9,25 +9,33 @@ function BookCard({ bookObj, onUpdate }) {
   // FOR DELETE, WE NEED TO REMOVE THE BOOK AND HAVE THE VIEW RERENDER,
   // SO WE PASS THE FUNCTION FROM THE PARENT THAT GETS THE BOOKS
   const deleteThisBook = () => {
-    if (window.confirm(`Delete ${bookObj.title}?`)) {
-      deleteBook(bookObj.firebaseKey).then(() => onUpdate());
+    if (bookObj && bookObj.title) {
+      if (window.confirm(`Delete ${bookObj.title}?`)) {
+        deleteBook(bookObj.firebaseKey).then(() => onUpdate());
+      }
+    } else {
+      console.error('Book or book title is undefined');
     }
   };
 
   return (
     <Card style={{ width: '18rem', margin: '10px' }}>
-      <Card.Img variant="top" src={bookObj.image} alt={bookObj.title} style={{ height: '400px' }} />
+      {bookObj && <Card.Img variant="top" src={bookObj.image} alt={bookObj.title} style={{ height: '400px' }} />}
       <Card.Body>
-        <Card.Title>{bookObj.title}</Card.Title>
-        <p className="card-text bold">{bookObj.sale && <span>SALE<br /></span> } ${bookObj.price}</p>
+        <Card.Title>{bookObj && bookObj.title}</Card.Title>
+        <p className="card-text bold">{bookObj && bookObj.sale && <span>SALE<br /></span> } ${bookObj && bookObj.price}</p>
         {/* DYNAMIC LINK TO VIEW THE BOOK DETAILS  */}
-        <Link href={`/book/${bookObj.firebaseKey}`} passHref>
-          <Button variant="primary" className="m-2">VIEW</Button>
-        </Link>
+        {bookObj && (
+          <Link href={`/book/${bookObj.firebaseKey}`} passHref>
+            <Button variant="primary" className="m-2">VIEW</Button>
+          </Link>
+        )}
         {/* DYNAMIC LINK TO EDIT THE BOOK DETAILS  */}
-        <Link href={`/book/edit/${bookObj.firebaseKey}`} passHref>
-          <Button variant="info">EDIT</Button>
-        </Link>
+        {bookObj && (
+          <Link href={`/book/edit/${bookObj.firebaseKey}`} passHref>
+            <Button variant="info">EDIT</Button>
+          </Link>
+        )}
         <Button variant="danger" onClick={deleteThisBook} className="m-2">
           DELETE
         </Button>
